requirements.txt
kivy
python-docx
markdown
plyer
pillow
